added a parsing utility and updated current CQL related tests to use it
