+++ cypress/fixtures/cqlSaveCQL.txt
@@ -1,0 +1,3 @@
+library TestMeasure version '0.0.014'
+using FHIR version '4.0.1'
+include FHIRHelpers version '4.0.001' called FHIRHelpers
